AM-1144: Drool rule unit tests: Moved classes to drools folder.
